Upgrade to use dehydrated
letsencrypt.sh was changed to dehydrated

This also causes us to change the paths we store files to, which
must be updated in existing containers using this image.

We maintain backward compatiblity, but older setups should migrate their
paths anyways as the backward compatibility might be dropped in the
future.

To migrate to this latest update:

- Change volume mounts:

  /opt/letsencrypt.sh/certs -> /opt/dehydrated/certs

  /opt/letsencrypt.sh/accounts -> /opt/dehydrated/accounts

- Update alias for acme directory:

  /var/www/letsencrypt -> /var/www/dehydrated

We also change to using Alpine version of nginx.